Rename tests to *.test.js
Matches jests default test pattern.
Is not an abbreviation/truncated word.
File responsibility is clear.
