+++ .npmignore
@@ -1,1 +1,1 @@
-**/*.specs.js
+**/*.test.js
